[PL-20981]: update list of accessible projects method (#28663)
* [PL-20981]: update list of accessible projects method

* changed the impl of projects count api

* incorporated project count change in aggregate service

* unit test fix

* checkstyle

* hash change

* hidden true

* hash change

* comments addressed

* comments addressed

* addressed comments

* clang format

* fixed unit test

+++ 960-ng-core-beans/src/main/java/io/harness/ng/core/dto/ActiveProjectsCountDTO.java
@@ -16,5 +16,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Schema(name = "ActiveProjectsCount", description = "This gives the count of the active Projects")
 public class ActiveProjectsCountDTO {
-  Integer count;
+  long count;
 }
